+++ src/plone/restapi/services/workflow/info.py
@@ -38,13 +38,14 @@ class WorkflowInfo(object):
             transitions.append({
                 '@id': '{}/@workflow/{}'.format(
                     self.context.absolute_url(), action['id']),
-                'title': action['title'],
+                'title': self.context.translate(action['title']),
             })
 
         for item, action in enumerate(history):
-            history[item]['title'] = wftool.getTitleForStateOnType(
-                action['review_state'],
-                self.context.portal_type)
+            history[item]['title'] = self.context.translate(
+                wftool.getTitleForStateOnType(
+                    action['review_state'],
+                    self.context.portal_type))
 
         result['workflow'].update({
             'history': json_compatible(history),